Step 0 : use ES6 modules and slice code a bit
