+++ src/xterm.js
@@ -31,25 +31,11 @@
  *   other features.
  */
 
-(function (xterm) {
-    if (typeof exports === 'object' && typeof module === 'object') {
-        /*
-         * CommonJS environment
-         */
-        module.exports = xterm.call(this);
-    } else if (typeof define == 'function') {
-        /*
-         * Require.js is available
-         */
-        define([], xterm.bind(window));
-    } else {
-        /*
-         * Plain browser environment
-         */
-        this.Xterm = xterm.call(this);
-        this.Terminal = this.Xterm; /* Backwards compatibility with term.js */
-    }
-})(function() {
+'use strict';
+var EventEmitter      = require('./lib/eventemitter');
+var CompositionHelper = require('./lib/compositionhelper');
+var Viewport          = require('./lib/viewport');
+
     /**
      * Terminal Emulation References:
      *   http://vt100.net/
@@ -61,7 +47,6 @@
      *   http://linux.die.net/man/7/urxvt
      */
 
-    'use strict';
 
     /**
      * Shared
@@ -69,393 +54,7 @@
 
     var window = this, document = this.document;
 
-    /**
-     * EventEmitter
-     */
 
-    function EventEmitter() {
-      this._events = this._events || {};
-    }
-
-    EventEmitter.prototype.addListener = function(type, listener) {
-      this._events[type] = this._events[type] || [];
-      this._events[type].push(listener);
-    };
-
-    EventEmitter.prototype.on = EventEmitter.prototype.addListener;
-
-    EventEmitter.prototype.removeListener = function(type, listener) {
-      if (!this._events[type]) return;
-
-      var obj = this._events[type]
-        , i = obj.length;
-
-      while (i--) {
-        if (obj[i] === listener || obj[i].listener === listener) {
-          obj.splice(i, 1);
-          return;
-        }
-      }
-    };
-
-    EventEmitter.prototype.off = EventEmitter.prototype.removeListener;
-
-    EventEmitter.prototype.removeAllListeners = function(type) {
-      if (this._events[type]) delete this._events[type];
-    };
-
-    EventEmitter.prototype.once = function(type, listener) {
-      var self = this;
-      function on() {
-        var args = Array.prototype.slice.call(arguments);
-        this.removeListener(type, on);
-        return listener.apply(this, args);
-      }
-      on.listener = listener;
-      return this.on(type, on);
-    };
-
-    EventEmitter.prototype.emit = function(type) {
-      if (!this._events[type]) return;
-
-      var args = Array.prototype.slice.call(arguments, 1)
-        , obj = this._events[type]
-        , l = obj.length
-        , i = 0;
-
-      for (; i < l; i++) {
-        obj[i].apply(this, args);
-      }
-    };
-
-    EventEmitter.prototype.listeners = function(type) {
-      return this._events[type] = this._events[type] || [];
-    };
-
-
-    /**
-     * Encapsulates the logic for handling compositionstart, compositionupdate and compositionend
-     * events, displaying the in-progress composition to the UI and forwarding the final composition
-     * to the handler.
-     * @param {HTMLTextAreaElement} textarea The textarea that xterm uses for input.
-     * @param {HTMLElement} compositionView The element to display the in-progress composition in.
-     * @param {Terminal} terminal The Terminal to forward the finished composition to.
-     */
-    function CompositionHelper(textarea, compositionView, terminal) {
-      this.textarea = textarea;
-      this.compositionView = compositionView;
-      this.terminal = terminal;
-
-      // Whether input composition is currently happening, eg. via a mobile keyboard, speech input
-      // or IME. This variable determines whether the compositionText should be displayed on the UI.
-      this.isComposing = false;
-
-      // The input currently being composed, eg. via a mobile keyboard, speech input or IME.
-      this.compositionText = null;
-
-      // The position within the input textarea's value of the current composition.
-      this.compositionPosition = { start: null, end: null };
-
-      // Whether a composition is in the process of being sent, setting this to false will cancel
-      // any in-progress composition.
-      this.isSendingComposition = false;
-    }
-
-    /**
-     * Handles the compositionstart event, activating the composition view.
-     */
-    CompositionHelper.prototype.compositionstart = function() {
-      this.isComposing = true;
-      this.compositionPosition.start = this.textarea.value.length;
-      this.compositionView.textContent = '';
-      this.compositionView.classList.add('active');
-    };
-
-    /**
-     * Handles the compositionupdate event, updating the composition view.
-     * @param {CompositionEvent} ev The event.
-     */
-    CompositionHelper.prototype.compositionupdate = function(ev) {
-      this.compositionView.textContent = ev.data;
-      this.updateCompositionElements();
-      var self = this;
-      setTimeout(function() {
-        self.compositionPosition.end = self.textarea.value.length;
-      }, 0);
-    };
-
-    /**
-     * Handles the compositionend event, hiding the composition view and sending the composition to
-     * the handler.
-     */
-    CompositionHelper.prototype.compositionend = function() {
-      this.finalizeComposition(true);
-    };
-
-    /**
-     * Handles the keydown event, routing any necessary events to the CompositionHelper functions.
-     * @return Whether the Terminal should continue processing the keydown event.
-     */
-    CompositionHelper.prototype.keydown = function(ev) {
-      if (this.isComposing || this.isSendingComposition) {
-        if (ev.keyCode === 229) {
-          // Continue composing if the keyCode is the "composition character"
-          return false;
-        } else if (ev.keyCode === 16 || ev.keyCode === 17 || ev.keyCode === 18) {
-          // Continue composing if the keyCode is a modifier key
-          return false;
-        } else {
-          // Finish composition immediately. This is mainly here for the case where enter is
-          // pressed and the handler needs to be triggered before the command is executed.
-          this.finalizeComposition(false);
-        }
-      }
-
-      if (ev.keyCode === 229) {
-        // If the "composition character" is used but gets to this point it means a non-composition
-        // character (eg. numbers and punctuation) was pressed when the IME was active.
-        this.handleAnyTextareaChanges();
-        return false;
-      }
-
-      return true;
-    };
-
-    /**
-     * Finalizes the composition, resuming regular input actions. This is called when a composition
-     * is ending.
-     * @param {boolean} waitForPropogation Whether to wait for events to propogate before sending
-     *   the input. This should be false if a non-composition keystroke is entered before the
-     *   compositionend event is triggered, such as enter, so that the composition is send before
-     *   the command is executed.
-     */
-    CompositionHelper.prototype.finalizeComposition = function(waitForPropogation) {
-      this.compositionView.classList.remove('active');
-      this.isComposing = false;
-      this.clearTextareaPosition();
-
-      if (!waitForPropogation) {
-        // Cancel any delayed composition send requests and send the input immediately.
-        this.isSendingComposition = false;
-        var input = this.textarea.value.substring(this.compositionPosition.start, this.compositionPosition.end);
-        this.terminal.handler(input);
-      } else {
-        // Make a deep copy of the composition position here as a new compositionstart event may
-        // fire before the setTimeout executes.
-        var currentCompositionPosition = {
-          start: this.compositionPosition.start,
-          end: this.compositionPosition.end,
-        }
-
-        // Since composition* events happen before the changes take place in the textarea on most
-        // browsers, use a setTimeout with 0ms time to allow the native compositionend event to
-        // complete. This ensures the correct character is retrieved, this solution was used
-        // because:
-        // - The compositionend event's data property is unreliable, at least on Chromium
-        // - The last compositionupdate event's data property does not always accurately describe
-        //   the character, a counter example being Korean where an ending consonsant can move to
-        //   the following character if the following input is a vowel.
-        var self = this;
-        this.isSendingComposition = true;
-        setTimeout(function () {
-          // Ensure that the input has not already been sent
-          if (self.isSendingComposition) {
-            self.isSendingComposition = false;
-            var input;
-            if (self.isComposing) {
-              // Use the end position to get the string if a new composition has started.
-              input = self.textarea.value.substring(currentCompositionPosition.start, currentCompositionPosition.end);
-            } else {
-              // Don't use the end position here in order to pick up any characters after the
-              // composition has finished, for example when typing a non-composition character
-              // (eg. 2) after a composition character.
-              input = self.textarea.value.substring(currentCompositionPosition.start);
-            }
-            self.terminal.handler(input);
-          }
-        }, 0);
-      }
-    };
-
-    /**
-     * Apply any changes made to the textarea after the current event chain is allowed to complete.
-     * This should be called when not currently composing but a keydown event with the "composition
-     * character" (229) is triggered, in order to allow non-composition text to be entered when an
-     * IME is active.
-     */
-    CompositionHelper.prototype.handleAnyTextareaChanges = function() {
-      var oldValue = this.textarea.value;
-      var self = this;
-      setTimeout(function() {
-        // Ignore if a composition has started since the timeout
-        if (!self.isComposing) {
-          var newValue = self.textarea.value;
-          var diff = newValue.replace(oldValue, '');
-          if (diff.length > 0) {
-            self.terminal.handler(diff);
-          }
-        }
-      }, 0);
-    };
-
-    /**
-     * Positions the composition view on top of the cursor and the textarea just below it (so the
-     * IME helper dialog is positioned correctly).
-     */
-    CompositionHelper.prototype.updateCompositionElements = function(dontRecurse) {
-      if (!this.isComposing) {
-        return;
-      }
-      var cursor = this.terminal.element.querySelector('.terminal-cursor');
-      if (cursor) {
-        this.compositionView.style.left = cursor.offsetLeft + 'px';
-        this.compositionView.style.top = cursor.offsetTop + 'px';
-        var compositionViewBounds = this.compositionView.getBoundingClientRect();
-        this.textarea.style.left = cursor.offsetLeft + compositionViewBounds.width + 'px';
-        this.textarea.style.top = (cursor.offsetTop + cursor.offsetHeight) + 'px';
-      }
-      if (!dontRecurse) {
-        setTimeout(this.updateCompositionElements.bind(this, true), 0);
-      }
-    };
-
-    /**
-     * Clears the textarea's position so that the cursor does not blink on IE.
-     * @private
-     */
-    CompositionHelper.prototype.clearTextareaPosition = function() {
-      this.textarea.style.left = '';
-      this.textarea.style.top = '';
-    };
-
-    /**
-     * Represents the viewport of a terminal, the visible area within the larger buffer of output.
-     * Logic for the virtual scroll bar is included in this object.
-     * @param {Terminal} terminal The Terminal object.
-     * @param {HTMLElement} viewportElement The DOM element acting as the viewport
-     * @param {HTMLElement} charMeasureElement A DOM element used to measure the character size of
-     *   the terminal.
-     */
-    function Viewport(terminal, viewportElement, scrollArea, charMeasureElement) {
-      this.terminal = terminal;
-      this.viewportElement = viewportElement;
-      this.scrollArea = scrollArea;
-      this.charMeasureElement = charMeasureElement;
-      this.currentRowHeight = 0;
-      this.lastRecordedBufferLength = 0;
-      this.lastRecordedViewportHeight = 0;
-
-      this.terminal.on('scroll', this.syncScrollArea.bind(this));
-      this.terminal.on('resize', this.syncScrollArea.bind(this));
-      this.viewportElement.addEventListener('scroll', this.onScroll.bind(this));
-
-      this.syncScrollArea();
-    }
-
-    /**
-     * Refreshes row height, setting line-height, viewport height and scroll area height if
-     * necessary.
-     * @param {number|undefined} charSize A character size measurement bounding rect object, if it
-     *   doesn't exist it will be created.
-     */
-    Viewport.prototype.refresh = function(charSize) {
-      var size = charSize || this.charMeasureElement.getBoundingClientRect();
-      if (size.height > 0) {
-        var rowHeightChanged = size.height !== this.currentRowHeight;
-        if (rowHeightChanged) {
-          this.currentRowHeight = size.height;
-          this.viewportElement.style.lineHeight = size.height + 'px';
-          this.terminal.rowContainer.style.lineHeight = size.height + 'px';
-        }
-        var viewportHeightChanged = this.lastRecordedViewportHeight !== this.terminal.rows;
-        if (rowHeightChanged || viewportHeightChanged) {
-          this.lastRecordedViewportHeight = this.terminal.rows;
-          this.viewportElement.style.height = size.height * this.terminal.rows + 'px';
-        }
-        this.scrollArea.style.height = (size.height * this.lastRecordedBufferLength) + 'px';
-      }
-    };
-
-    /**
-     * Updates dimensions and synchronizes the scroll area if necessary.
-     */
-    Viewport.prototype.syncScrollArea = function() {
-      if (this.isApplicationMode) {
-        // Fix scroll bar in application mode
-        this.lastRecordedBufferLength = this.terminal.rows;
-        this.refresh();
-        return;
-      }
-
-      if (this.lastRecordedBufferLength !== this.terminal.lines.length) {
-        // If buffer height changed
-        this.lastRecordedBufferLength = this.terminal.lines.length;
-        this.refresh();
-      } else if (this.lastRecordedViewportHeight !== this.terminal.rows) {
-        // If viewport height changed
-        this.refresh();
-      } else {
-        // If size has changed, refresh viewport
-        var size = this.charMeasureElement.getBoundingClientRect();
-        if (size.height !== this.currentRowHeight) {
-          this.refresh(size);
-        }
-      }
-
-      // Sync scrollTop
-      var scrollTop = this.terminal.ydisp * this.currentRowHeight;
-      if (this.viewportElement.scrollTop !== scrollTop) {
-        this.viewportElement.scrollTop = scrollTop;
-      }
-    };
-
-    /**
-     * Sets the application mode of the viewport.
-     * @param {boolean} isApplicationMode Sets whether the terminal is in application mode. true
-     * for application mode (DECKPAM) and false for normal mode (DECKPNM).
-     */
-    Viewport.prototype.setApplicationMode = function(isApplicationMode) {
-      this.isApplicationMode = isApplicationMode;
-      this.syncScrollArea();
-    };
-
-    /**
-     * Handles scroll events on the viewport, calculating the new viewport and requesting the
-     * terminal to scroll to it.
-     * @param {Event} ev The scroll event.
-     */
-    Viewport.prototype.onScroll = function(ev) {
-      if (this.isApplicationMode) {
-        // Scrolling via the scroll bar is disabled during application mode
-        return;
-      }
-      var newRow = Math.round(this.viewportElement.scrollTop / this.currentRowHeight);
-      var diff = newRow - this.terminal.ydisp;
-      this.terminal.scrollDisp(diff, true);
-    };
-
-    /**
-     * Handles mouse wheel events by adjusting the viewport's scrollTop and delegating the actual
-     * scrolling to `onScroll`, this event needs to be attached manually by the consumer of
-     * `Viewport`.
-     * @param {WheelEvent} ev The mouse wheel event.
-     */
-    Viewport.prototype.onWheel = function(ev) {
-      if (ev.deltaY === 0) {
-        // Do nothing if it's not a vertical scroll event
-        return;
-      }
-      // Fallback to WheelEvent.DOM_DELTA_PIXEL
-      var multiplier = 1;
-      if (ev.deltaMode === WheelEvent.DOM_DELTA_LINE) {
-        multiplier = this.currentRowHeight;
-      } else if (ev.deltaMode === WheelEvent.DOM_DELTA_PAGE) {
-        multiplier = this.currentRowHeight * this.terminal.rows;
-      }
-      this.viewportElement.scrollTop += ev.deltaY * multiplier;
-      // Prevent the page from scrolling when the terminal scrolls
-      ev.preventDefault();
-    };
 
     /**
      * States
@@ -5445,5 +5044,4 @@
     Terminal.cancel = cancel;
 
 
-    return Terminal;
-});
+module.exports = Terminal;
